+++ src/pymonzo/whoami/schemas.py
@@ -2,6 +2,14 @@
 
 from pydantic import BaseModel
 
+# Optional `rich` support
+try:
+    from rich.table import Table
+except ImportError:
+    RICH_AVAILABLE = False
+else:
+    RICH_AVAILABLE = True
+
 
 class MonzoWhoAmI(BaseModel):
     """API schema for a 'whoami' object.
@@ -18,3 +26,16 @@ class MonzoWhoAmI(BaseModel):
     authenticated: bool
     client_id: str
     user_id: str
+
+    if RICH_AVAILABLE:
+
+        def __rich__(self) -> Table:
+            """Pretty printing support for `rich`."""
+            grid = Table.grid(padding=(0, 5))
+            grid.add_column(style="bold yellow")
+            grid.add_column()
+            grid.add_row("Authenticated:", "Yes" if self.authenticated else "No")
+            grid.add_row("Client ID:", self.client_id)
+            grid.add_row("User ID:", self.user_id)
+
+            return grid